avoid segfault when watching an eden subdir
Summary:
When performing "watchman watch eden/subdir" we would crash rather than give a meaningful message.

This diff corrects the crash and helps to avoid a similar situation in the
future by promoting a nullptr return to an exception with a message explaining
what should be done to make this work properly.

I've also introduced an exception type that can be used to prevent the "auto"
watch resolver from falling back to a lower priority watcher implementation; in
this specific circumstance (watching a subdir rather than a root) we don't want
to fall back to inotify as we don't want to allow watchman to perform an
O(almost-whole-repo) crawl.

Reviewed By: eamonnkent

Differential Revision: D5598374

fbshipit-source-id: 18522f5ce33cc0394bcc6724b221464f91fe8d1f