Force the path generator for scm queries with changed merge base
Summary: In scm aware watchman queries where the merge base has changed, watchman uses a path generator.  This generator is applied to the list of files returned by a query the query scm->getFilesChangedSinceMergeWithBase().  In the case where this file list was empty, the wrong generator was chosen.  The change is to use a lambda to force the setting of the generator (rather than relying on generator lookup based on query).

Reviewed By: wez

Differential Revision: D5645290

fbshipit-source-id: b346da4f4c898f537523d5c60dcfcdf8b847fe08