Support query expression aggregation
Summary:
We are adding support to suffix expressions that allow queries to pass a list
suffixes to match. Alternatively, users could use an 'anyof' expression and
a list of suffix expressions. This commit is an optimization to convert an
any of list of suffix expressions into a single suffix expression (which
provides better performance)

Reviewed By: wez

Differential Revision: D5662260

fbshipit-source-id: c6dd460a97a3744df0fa3019c417e162fb9e5490