Refactor resolve_root_or_err to throw exceptions
Summary:
In the current form of resolve_root_or_err,  its behavior is:
1) returns a string error message
2) returns null on failure
3) notifies client connection of failure
All clients must handle the null return - which can lead to bugs.

We updated the implementation to be less error prone:
1) throw an exception on error - that contains the error message.
This frees clients from having to null check (command dispatch
will take care of catching the exception).

Note: we still have to pass in the client to the resolve_root method because
there are checks for client_is_owner and client_mode.

Reviewed By: wez

Differential Revision: D5826757

fbshipit-source-id: 579f55c64ca9daf3fa4ebd4a6ac4d7dab19a2d53