Fix circular dependency in debug build
Summary:
The debug build is broken because of a circular dependency. watchman_string
uses w_assert, which, in turn introduces a dependency back to watchman_string.
It only occurs in the debug build (assertion).

The fix is to throw a runtime_error. (Alternatively, we could just do a plane
assert).

Reviewed By: wez

Differential Revision: D5933196

fbshipit-source-id: 5eaf26578d7cf2c4a868793872fdbc2c0cbf1b5a