don't fallback to inotify when eden fails to respond
Summary:
bolinfest runs into a situation where we've wound up using
the inotify watcher inside eden and that's pretty awful because
inotify is inherently `O(repo)`.

This diff causes any errors that manifest after we have resolved
the root of the eden repo to be terminal wrt. resolver a watcher
implementation.

Reviewed By: eamonnkent

Differential Revision: D6143177

fbshipit-source-id: bbfc559679f175bc592832ae621150d9bbadf15d