avoid test breakage by unmounting eden at the end
Summary:
This is in the "shouldn't be a problem" category.  It appears
that eden gets blocked waiting for thrift worker threads to shutdown
when we ask it to shutdown in this test environment.  This manifests
as the `eden shutdown` command resorting to kill and bubbling up a
failure in the test cleanup.

Inspecting the eden process with gdb shows that none of those threads are
still around.  Even weirder, just attaching and detaching gdb seems to
knock the process out of its sleep state and causes the test to pass.

In the interest of reducing noise for D6310325 we can workaround this
here by explicitly unmounting before we get to the cleanup.

I don't like this very much, but I think it's a reasonable thing to do.

Separately, I think we should run down the root cause of this apparent
deadlock on shutdown as I think it is the root cause of some of the
problem reports in our dogfooding group.

Reviewed By: bolinfest

Differential Revision: D6343496

fbshipit-source-id: 6d4e8ce0d83b1b2db998518bc592c14435f48beb