enable retrying for nodejs tests
Summary:
I observed that we're not kicking in with the flaky test retry
machinery and that this is causing travis to report some builds
as failed.

This diff enables retrying, and also allows for picking up `nodejs`
from the PATH which is how my Ubuntu 16.04 system installs the node
binary.
Closes https://github.com/facebook/watchman/pull/560

Differential Revision: D6622884

Pulled By: wez

fbshipit-source-id: 2294dc45eb6f5d0ccdb0ec4a1135396d470b70bd