Add --no-site-spawner option.
Summary:
Adds a `--no-site-spawner` option that bypasses both the site-specific spawn hook and any built-in custom/platform-specific spawning functionality, using the built-in basic `daemonize()` spawner.

It may feel like an over-reach to have this bypass launchd, but one of the things that wez would like to move to in the future is having the launchd spawning logic become a site-specific launch hook that is configured and deployed as part of the homebrew packaging.

This option is not expected to be used in many use cases.

Fixes #558
Closes https://github.com/facebook/watchman/pull/559

Reviewed By: simpkins

Differential Revision: D6660690

Pulled By: wez

fbshipit-source-id: 4f87390291a8509b689b7dc0a9d12cc04d95828d