improve error handling in getattrlistbulk()
Summary:
We've been seeing some issues that seem to correlate with
users on APFS, so we're overdue revisiting our usage of this
complex syscall.

We're using `getattrlistbulk(2)` as a way to batch together `opendir(3)` and a
series of `lstat(2)` calls.  This syscall uses an ABI defined in
`getattrlist(2)`.

Based on some problem reports, it seems that APFS is more prone to encountering
error conditions or that it doesn't report all of the information that we
request all of the time.

This diff improves our handling of this syscall:

1. Use the `returned` field to examine which of the requested fields were
   returned before trying to use them.  This matters the most for the `err`
   field which we seen logging stack garbage in place of its usual `errno`
   values.
2. Make a better effort at falling back to the `opendir(3)` mode of operation
   if we got the name of the entry but no other stat info.

Reviewed By: simpkins

Differential Revision: D6991054

fbshipit-source-id: ccb1da67f76d9c7f253a52819e7b12c291855ee0