avoid pessimistic recursive crawl on linux
Summary:
We had the somewhat unfortunate behavior of initiating a
recursive crawl of the parent directory of a removed directory
entry on linux.

This means that in a big repo `mkdir foo ; rmdir foo` at the root
causes the entire repo to be recrawled and block concurrent access
(eg: `hg status`) while waiting for that to happen.

This diff removes the recursive flag; this is safe to do because
we'll look at the parent and enumerate its entries.  Our normal
behavior is to then recurse into a directory if it is new, and
to observe the deletion of the named entry otherwise.

Our behavior around this is covered by our baseline behavior
integration tests.

Reviewed By: chadaustin

Differential Revision: D7006121

fbshipit-source-id: 1b19ee8adde5b14e6253cb5296f7e5c3051fcfd7