avoid `abort` in a number of code paths
Summary:
We see this manifest most often in production in environments
that are nonsensical; for example, spawned by a client process whose
executable image has been deleted and who is running with a cwd that
has been deleted.

We'll abort when run in such an environment because we're not able
to spawn watchman because the site spawner script that has been
configured is a python script that fails to resolve the cwd and
exits.

These aborts are captured and reported across our production fleet
and make us look bad in the coredumper data set.

Rather than abort let's just `exit(1)`.

This diff replaces a number of `abort()` calls with `log(FATAL,...`)
and removes a hangover from our C/malloc days where we were left
checking the result of `new` against `nullptr`; this should always
throw `bad_alloc` and thus the conditional should now just be
deadcode.

Reviewed By: simpkins

Differential Revision: D7060727

fbshipit-source-id: 57483cae023631de899e3c0c04795abf259f6f85