reduce flakiness of test_wm_wait.py
Summary:
A couple of adjustments to make this happier on the open source CI.

We don't need to check that the parent dirs of the changed items show up,
as whether that happens or not is OS dependent.  We therefore check to
make sure that we at least see the changed items in there, and increase
the grace period (the `-t` option) and the number of events (`-m` option)
to accomodate some variance.

Reviewed By: simpkins

Differential Revision: D7106677

fbshipit-source-id: 48fb3b751f75d16a0a6a169de79b543f67367722