fix existential issue with SCM aware queries
Summary:
This was a bit of a thorny issue to run down.  What was happening was
this: when source control aware mode is enabled we subvert the normal recency
index and use the `pathGenerator` rather than the `timeGenerator`.

We pass the list of files that we got back from the source control system to
the path generator and let it evaluate the query based on information that
we've observed in the InMemoryView.

If watchman has never observed some components of the paths returned by the
source control query, such as in the case where watchman was restarted after
checking out the revision, and we're asked about changes since an earlier hash,
the `pathGenerator` will not be able to furnish information about the paths
returned by source control.   This means that we are potentially inaccurate for
files that exist on the filesystem (although we will likely soon observe the
current state of those files), but more critically, we cannot return
information about files that were deleted in between the commits in the query
range.

The resolution is not to use the InMemoryView as the source of information
about these files.

This diff adds a LocalFileResult for just this purpose; this is an
implementation of the FileResult class that knows how to stat the local
filesystem as needed to furnish the metadata that may be needed by the query.
I'm not thrilled that it needs to stat the files, but this the simplest way to
obtain correct results and we can figure out if we need to do better than this
in practice as a follow up.

I've also broken out the guts of `ContentHashCache::computeHashImmediate`
as a static function so that we can call it from `LocalFileResult`.
Ideally we'd find a way to plumb in the cache that is held in the
`InMemoryView`, but we're operating at a different layer and cannot
guarantee that it is present; it isn't present for Eden and we still
need to solve the hash computation issue for Eden.  Those are out of
scope wrt. resolving this issue in the short term.

Reviewed By: simpkins

Differential Revision: D7142088

fbshipit-source-id: 35b7d7745e30adabb8b9a7b3713d6ba3500b3552