speculative workaround btrfs inotify issue
Summary:
This is follow on from the same root issue as the one tackled in D7014364.

In this flavor of the problem, on a btrfs volume, the repo has been removed
and replaced by a different repo.  Since the repo was replaced, writing cookies
to either the cookie dir or the root will not elicit inotify events and we
can never sync up for queries.

This diff attempts to tackle this issue by checking the inode numbers on
directories.

1. Previously we would use the inode number change to detect if a file had changed
but we didn't force a recursive examination of the node if it happened to
be a directory.  This diff adds forcing on the recursive flag in that case.

2. We would never `stat()` the root inode after starting up because we don't
maintain data structures that allow the root path to work with the
`InMemoryView::statPath` method.  Consequently we need to explicitly add an
inode number check to the `crawler` method that is used to handle directory
contents.  We trigger a recrawl in this situation (rather than force on a
recursive flag) because the inotify watches may hold on to the now invisible
inodes for the removed subvolume indefinitely (that was the main complaint in
D7014364) so it is best to throw away our inotify instance and set up a new
one.

3. We need to trigger the above checks so we wire that in to the cookie sync
error path.  In practice this will happen in response to a timeout and the IO
will be scheduled async; detection and remediation will happen asynchronously
wrt. the current query.

We don't have a way to detect this problem in a timely fashion around the time
of the replacement; this diff is focused on detecting and unbreaking things
around the time of a subsequent query.  It will happen too late for that query
to succeed but should at least get us into a more consistent state.

Reviewed By: simpkins

Differential Revision: D7250428

fbshipit-source-id: 3f261e06c8b38f4d063ec9380c617f5d22eead78